requirements.txt
discord.py[voice]==2.3.2
gTTS==2.4.0
PyNaCl==1.5.0
python-dotenv==1.0.0
openai==1.12.0 